fix: Action toolbar menu state (#38964)
