+++ app/client/src/PluginActionEditor/components/PluginActionContextMenu/index.ts
@@ -1,0 +1,1 @@
+export { PluginActionContextMenu } from "./PluginActionContextMenu";
